# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Gestion Magasin Documentation",
        "description": "Implementation solution de Gestion de Magasin avec Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/commandeachats": {
            "get": {
                "tags": [
                    "Affiche Commandeachats et Lignecommandeachats"
                ],
                "summary": "Affiche la liste des Commandeachats et Lignecommandeachats",
                "description": "Retourner Touts les Commandeachats et Lignecommandeachats",
                "operationId": "Affiche Commandeachats et Lignecommandeachats",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ajouter Commandeachats et Lignecommandeachats"
                ],
                "summary": "Ajoute Commandeachats et Lignecommandeachats",
                "description": "Ajouter  Commandeachats et Lignecommandeachats et Retourne les Valeurs Ajouter",
                "operationId": "Store Commandeachats et Lignecommandeachats",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/commandeachats/{commandeachat}": {
            "get": {
                "tags": [
                    "Affiche Commandeachat et Lignecommandeachat par ID"
                ],
                "summary": "Affiche le Commandeachat et Lignecommandeachat par ID",
                "description": "Retourner le Commandeachat et Lignecommandeachat",
                "operationId": "Show Commandeachat et Lignecommandeachat",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Mise en Jour Commandeachat et Lignecommandeachat"
                ],
                "summary": "Modifier Commandeachat et Lignecommandeachat",
                "description": "Retourner Commandeachat et Lignecommandeachat Modifier",
                "operationId": "Modifier Commandeachat et Lignecommandeachat",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Supprimer Commandeachat et Lignecommandeachat"
                ],
                "summary": "Supprimer un Commandeachat et Lignecommandeachat",
                "description": "Supprimer un Commandeachat et Lignecommandeachat",
                "operationId": "Delete Commandeachat et Lignecommandeachat",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    }
}